Keep track of all vars in a vector
This fixes the `forEachVar` function which up to this point only relied on the list of short names. However, as variables may have only a short or a long name, iteration is more complicated when each var is supposed to be visited only once. Hence a simple vector of var pointers is added to keep track of the variables currently registered.